nodeSorter sorts by original array position, not by name
If two nodes had the same score, previously, the nodeSorter would
compare by names of the nodes.

With this change, instead the nodeSorter compares by the position of
the nodes in the input nodesAll array.

This allows the application to have more control and flexibility,
where to have the same behavior as before, the application should just
pass in a nodesAll array sorted by node names.

With this change, an advanced application can now pass in a nodesAll
array sorted by how blance should favor the nodes in planning, which
some applications may use to shuffle or rotate the nodesAll amongst
resources within the same cluster to achieve better utilization.

Change-Id: I4f068e3940cc671c5133a32d59d430f7459f2686
Reviewed-on: http://review.couchbase.org/57487
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>